requirements.txt
amqp==5.3.1
asgiref==3.8.1
astroid==3.3.11
billiard==4.2.1
celery==5.5.3
celery-stubs==0.1.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
<<<<<<< HEAD
cron_descriptor==2.0.6
dill==0.4.0
Django==5.2.3
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-filter==25.1
django-silk==5.4.2
=======
cron-descriptor==1.4.5
dill==0.4.0
Django==5.2.3
django-celery-beat==2.8.1
django-debug-toolbar==5.2.0
django-redis==6.0.0
>>>>>>> eb7e6f33adc0c1c5f46302bc3c5bab451a16e0fb
django-stubs==5.2.1
django-stubs-ext==5.2.1
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
<<<<<<< HEAD
gprof2dot==2025.4.14
=======
>>>>>>> eb7e6f33adc0c1c5f46302bc3c5bab451a16e0fb
isort==6.0.1
kombu==5.5.4
mccabe==0.7.0
packaging==25.0
pillow==11.3.0
<<<<<<< HEAD
platformdirs==4.4.0
=======
platformdirs==4.3.8
>>>>>>> eb7e6f33adc0c1c5f46302bc3c5bab451a16e0fb
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg2-binary==2.9.10
PyJWT==2.9.0
<<<<<<< HEAD
pylint==3.3.8
=======
>>>>>>> eb7e6f33adc0c1c5f46302bc3c5bab451a16e0fb
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.2.0
setuptools==80.9.0
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
tomlkit==0.13.3
types-PyYAML==6.0.12.20250516
typing_extensions==4.14.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
django-filter~=25.1